requirements.txt
jinja2
jira
requests
pyyaml
redis
six

git+https://github.com/ianunruh/pyzendesk.git
